+++ jam/data/HistDouble2D.java
@@ -77,8 +77,8 @@ final class HistDouble2D extends AbstractHist2D {
 		return counts2dD;
 	}
 
-	public int getCount() {
-		return (int) getArea();
+	public double getCount() {
+		return getArea();
 	}
 	/* (non-Javadoc)
 	 * @see jam.data.Histogram#setZero()
